Major refactor & bug fixes
Added displayable error codes when the rom you try to read in fails in some way.

Both main thread and window thread come to a close when either the window closes or the user presses escape.

Sped up the GetKeyState function= in bus.cpp by adding return statements in each IF.

Added new functions to System which allows the user to start the system with one function call.

Wrote a class which allows you to easily create a program which displays a string of text comprised of all sprites in the default fontset.

Changed the keys from the exact keyboard equivilants (e.g. F in the game is F on the kb, & 9 in the game is 9 on the kb). To the 4x4 set of keys beginning at 1 on the kb, and ending in in v on the kb.